Added function for temporal complexity reduction

+++ edisgo/flex_opt/check_tech_constraints.py
@@ -1002,8 +1002,8 @@ def _lv_allowed_voltage_limits(edisgo_obj, lv_grids=None, mode=None):
         config_string = "mv_lv_station"
 
         # get all primary and secondary sides
-        primary_sides = pd.Series()
-        secondary_sides = pd.Series()
+        primary_sides = pd.Series(dtype="float64")
+        secondary_sides = pd.Series(dtype="float64")
         for grid in lv_grids:
             primary_side = grid.transformers_df.iloc[0].bus0
             if primary_side in buses_in_pfa:
@@ -1037,7 +1037,7 @@ def _lv_allowed_voltage_limits(edisgo_obj, lv_grids=None, mode=None):
 
         # get all secondary sides and buses in grids
         buses_dict = {}
-        secondary_sides = pd.Series()
+        secondary_sides = pd.Series(dtype="float64")
         for grid in lv_grids:
             secondary_side = grid.station.index[0]
             if secondary_side in buses_in_pfa:
